move reused functions to helper files
