+++ src/components/integrations/AuthWidget.tsx
@@ -49,9 +49,9 @@ import {
 } from "@/integrations/autoConfigure";
 import { freshIdentifier } from "@/utils/variableUtils";
 import { selectIntegrationConfigs } from "@/integrations/store/integrationsSelectors";
-import { type Schema } from "@/types/schemaTypes";
 import useAsyncState from "@/hooks/useAsyncState";
 import AsyncStateGate from "@/components/AsyncStateGate";
+import { convertSchemaToConfigState } from "@/components/integrations/integrationHelpers";
 
 const { upsertIntegrationConfig, deleteIntegrationConfig } =
   integrationsSlice.actions;
@@ -82,63 +82,6 @@ const RefreshButton: React.VFC<{
     </Button>
   );
 };
-
-function isRequired(schema: Schema, key: string): boolean {
-  return schema.required && schema.required.includes(key);
-}
-
-export function convertSchemaToConfigState(inputSchema: Schema): UnknownObject {
-  const result: UnknownObject = {};
-  for (const [key, value] of Object.entries(inputSchema.properties)) {
-    if (
-      typeof value === "boolean" ||
-      value.type === "null" ||
-      !isRequired(inputSchema, key)
-    ) {
-      continue;
-    }
-
-    if (value.type === "object") {
-      // eslint-disable-next-line security/detect-object-injection -- Schema property keys
-      result[key] = convertSchemaToConfigState(value);
-    } else {
-      if (value.default !== undefined) {
-        // eslint-disable-next-line security/detect-object-injection -- Schema property keys
-        result[key] = value.default;
-        continue;
-      }
-
-      switch (value.type) {
-        case "boolean": {
-          // eslint-disable-next-line security/detect-object-injection -- Schema property keys
-          result[key] = false;
-          break;
-        }
-
-        case "number":
-        case "integer": {
-          // eslint-disable-next-line security/detect-object-injection -- Schema property keys
-          result[key] = 0;
-          break;
-        }
-
-        case "array": {
-          // eslint-disable-next-line security/detect-object-injection -- Schema property keys
-          result[key] = [];
-          break;
-        }
-
-        default: {
-          // eslint-disable-next-line security/detect-object-injection -- Schema property keys
-          result[key] = "";
-          break;
-        }
-      }
-    }
-  }
-
-  return result;
-}
 
 type AuthWidgetContentProps = {
   name: string;
